+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 
 ## Changelog
 
-#### [Upcoming](https://github.com/kuvaus/coffeepy/compare/v0.1.7...HEAD)
+#### [Upcoming](https://github.com/kuvaus/coffeepy/compare/v0.1.8...HEAD)
+
+#### [v0.1.8](https://github.com/kuvaus/coffeepy/releases/tag/v0.1.8)
 
 - Fix `--help` on windows systems  without unicode support 
 - Better error handling on Linux